move minileaderboard to its only use
